+++ src/NNHelferlein.jl
@@ -56,6 +56,7 @@ export DNN, Classifier, Regressor, Chain, VAE,          # chains
        Embed, Recurrent,
        Softmax, Dropout, BatchNorm, LayerNorm,
        Linear, GaussianNoise,
+       GlobalAveragePooling,
        get_hidden_states, get_cell_states,
        set_hidden_states!, set_cell_states!,
        reset_hidden_states!, reset_cell_states!,