+++ docs/src/changelog.md
@@ -1,5 +1,8 @@
 # ChangeLog of NNHelferlein package
 
++ docu fixed for DeConv
++ changelog added to docu
+
 ### v1.1.0
 + documentation for release added
 + split_minibatches() made stable (never returns an empty iterator)