+++ docs/src/changelog.md
@@ -1,11 +1,12 @@
 # ChangeLog of NNHelferlein package
 
+### v1.1.1
++ some docstring cosmetics
 + Activation Layers added
 + layer GlobalAveragePoling added
 + pre-trained vgg example fixed for changed "import-HDF"-interface
 + hdf5 import with all kwargs possible
 + added: Layer + Layer = Chain
-+ docu fixed for DeConv
 + changelog added to docu
 
 ### v1.1.0